moving lens_distoriton back to ceres
