+++ arrows/mvg/camera_options.cxx
@@ -8,7 +8,8 @@
  */
 
 #include "camera_options.h"
-#include "lens_distortion.h"
+
+#include "../ceres/lens_distortion.h"
 
 #define MVG_ENUM_HELPERS( NS, mvg_type ) \
   namespace kwiver { \
@@ -103,23 +104,6 @@ StringToLensDistortionType( std::string value, LensDistortionType* type )
   STRENUM( POLYNOMIAL_RADIAL_TANGENTIAL_DISTORTION );
   STRENUM( RATIONAL_RADIAL_TANGENTIAL_DISTORTION );
   return false;
-}
-
-/// Return the number of distortion parameters required for each type
-unsigned int
-num_distortion_params( LensDistortionType type )
-{
-  switch( type )
-  {
-    case POLYNOMIAL_RADIAL_DISTORTION:
-      return distortion_poly_radial::num_coeffs;
-    case POLYNOMIAL_RADIAL_TANGENTIAL_DISTORTION:
-      return distortion_poly_radial_tangential::num_coeffs;
-    case RATIONAL_RADIAL_TANGENTIAL_DISTORTION:
-      return distortion_ratpoly_radial_tangential::num_coeffs;
-    default:
-      return 0;
-  }
 }
 
 /// Constructor
